+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed binding target resolution infinite loop
+
 ## v0.2.7 (2022-11-26)
 * Added key-value store to container
 * Added referential type aliases to bound objects